fixed some issues. Placement now take account of feature placement distance

+++ src/main/java/net/rodofire/easierworldcreator/mixin/world/gen/ChunkGeneratorMixin.java
@@ -1,15 +1,18 @@
 package net.rodofire.easierworldcreator.mixin.world.gen;
 
 import com.llamalad7.mixinextras.sugar.Local;
+import net.minecraft.util.math.ChunkPos;
 import net.minecraft.world.StructureWorldAccess;
 import net.minecraft.world.chunk.Chunk;
 import net.minecraft.world.gen.GenerationStep;
 import net.minecraft.world.gen.StructureAccessor;
 import net.minecraft.world.gen.chunk.ChunkGenerator;
 import net.minecraft.world.gen.feature.PlacedFeature;
+import net.rodofire.easierworldcreator.blockdata.blocklist.BlockListHelper;
 import net.rodofire.easierworldcreator.blockdata.blocklist.BlockListManager;
 import net.rodofire.easierworldcreator.shape.block.placer.WGShapeHandler;
 import net.rodofire.easierworldcreator.shape.block.placer.WGShapePlacerManager;
+import net.rodofire.easierworldcreator.util.file.EwcFolderData;
 import net.rodofire.easierworldcreator.util.file.FileUtil;
 import net.rodofire.easierworldcreator.util.file.LoadChunkShapeInfo;
 import org.spongepowered.asm.mixin.Mixin;
@@ -18,17 +21,20 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
+import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.List;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.stream.Stream;
 
 /**
  * mixin to change how are the chunk generated to include multi-chunk features
  */
 @Mixin(ChunkGenerator.class)
 public abstract class ChunkGeneratorMixin {
-
     @Unique
-    WGShapePlacerManager placerManager;
+    WGShapePlacerManager[] placerManagers;
 
     @Unique
     PlacedFeature old;
@@ -38,22 +44,35 @@ public abstract class ChunkGeneratorMixin {
      */
     @Inject(method = "generateFeatures", at = @At(value = "HEAD"))
     private void initShapeHandler(StructureWorldAccess world, Chunk chunk, StructureAccessor structureAccessor, CallbackInfo ci) {
-        placerManager = WGShapeHandler.decodeInformation(chunk.getPos());
+        Set<WGShapePlacerManager> shapePlacer = new HashSet<>();
+        for (int i = -1; i <= 1; i++) {
+            for (int j = -1; j <= 1; j++) {
+                ChunkPos pos = new ChunkPos(chunk.getPos().x + i, chunk.getPos().z + j);
+                shapePlacer.add(WGShapeHandler.decodeInformation(pos));
+            }
+        }
+
+        placerManagers = shapePlacer.toArray(new WGShapePlacerManager[0]);
     }
 
 
     @Inject(method = "generateFeatures", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/gen/StructureAccessor;shouldGenerateStructures()Z"))
-    private void onGenerationStep(StructureWorldAccess world, Chunk chunk, StructureAccessor structureAccessor, CallbackInfo ci, @Local(ordinal = 0) int k) {
-        if (placerManager == null) return;
-        if (k >= GenerationStep.Feature.values().length) {
-            return;
-        }
-        Path[] paths = placerManager.getToPlace(GenerationStep.Feature.values()[k]);
-        for (Path path : paths) {
-            world.setCurrentlyGeneratingStructureName(() -> "ewc multi-chunk feature generating: " + path.getFileName());
-            BlockListManager comparator = LoadChunkShapeInfo.loadFromJson(world, path);
-            LoadChunkShapeInfo.placeStructure(world, comparator);
-            FileUtil.removeFile(path);
+    private void onGenerationStep(StructureWorldAccess world, Chunk chunk, StructureAccessor structureAccessor, CallbackInfo ci, @Local(ordinal = 2) int k) {
+        for (WGShapePlacerManager placerManager : placerManagers) {
+            if (placerManager == null) return;
+
+            if (k >= GenerationStep.Feature.values().length) {
+                return;
+            }
+
+            Path[] paths = placerManager.getToPlace(GenerationStep.Feature.values()[k]);
+            for (Path path : paths) {
+                world.setCurrentlyGeneratingStructureName(() -> "\n\t-ewc multi-chunk feature generating: \n\t\t- " + path.getFileName() + "\n\t\t - step : generation step");
+                BlockListManager manager = BlockListHelper.fromJsonPath(world, path);
+                if (manager == null) continue;
+                manager.placeAllNDelete(world);
+                FileUtil.removeFile(path);
+            }
         }
     }
 
@@ -73,13 +92,15 @@ public abstract class ChunkGeneratorMixin {
      */
     @Inject(method = "generateFeatures", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/StructureWorldAccess;setCurrentlyGeneratingStructureName(Ljava/util/function/Supplier;)V", ordinal = 1))
     private void onFeatureGenerated(StructureWorldAccess world, Chunk chunk, StructureAccessor structureAccessor, CallbackInfo ci, @Local PlacedFeature placedFeature) {
-        if (placerManager == null) return;
-        Path[] paths = placerManager.getToPlace(old, placedFeature);
-        for (Path path : paths) {
-            world.setCurrentlyGeneratingStructureName(() -> "ewc multi-chunk feature generating: " + path.getFileName());
-            BlockListManager comparator = LoadChunkShapeInfo.loadFromJson(world, path);
-            LoadChunkShapeInfo.placeStructure(world, comparator);
-            FileUtil.removeFile(path);
+        for (WGShapePlacerManager placerManager : placerManagers) {
+            if (placerManager == null) return;
+            Path[] paths = placerManager.getToPlace(old, placedFeature);
+            for (Path path : paths) {
+                world.setCurrentlyGeneratingStructureName(() -> "\n\t-ewc multi-chunk feature generating: \n\t\t- " + path.getFileName() + "\n\t\t - step : feature");
+                BlockListManager comparator = LoadChunkShapeInfo.loadFromJson(world, path);
+                LoadChunkShapeInfo.placeStructure(world, comparator);
+                FileUtil.removeFile(path);
+            }
         }
 
         old = placedFeature;
@@ -87,7 +108,8 @@ public abstract class ChunkGeneratorMixin {
 
     /**
      * At then end, with the fact that multi chunk are multithreaded, some files might not have been placed,
-     * that's why we add this mixin at the end of the method to make sure that all the files were placed
+     * that's why we add this mixin at the end of the method to make sure that all the files were placed.
+     * We also remove all files related to this chunk generation
      *
      * @param world             the world of the chunk
      * @param chunk             the chunk generated
@@ -96,16 +118,43 @@ public abstract class ChunkGeneratorMixin {
      */
     @Inject(method = "generateFeatures", at = @At(value = "TAIL"))
     private void endGeneration(StructureWorldAccess world, Chunk chunk, StructureAccessor structureAccessor, CallbackInfo ci) {
-        List<Path> pathlist = LoadChunkShapeInfo.getWorldGenFiles(world, chunk);
-        if (!pathlist.isEmpty()) {
-            for (Path path : pathlist) {
-                world.setCurrentlyGeneratingStructureName(() -> "ewc multi-chunk feature generating: " + path.getFileName());
-                BlockListManager comparator = LoadChunkShapeInfo.loadFromJson(world, path);
-                LoadChunkShapeInfo.placeStructure(world, comparator);
-                FileUtil.removeFile(path);
+        for (int i = -1; i <= 1; i++) {
+            for (int j = -1; j <= 1; j++) {
+                ChunkPos pos = new ChunkPos(chunk.getPos().x + i, chunk.getPos().z + j);
+                Path basePaths = EwcFolderData.getStructureDataDir(pos);
+                try (Stream<Path> paths = Files.list(basePaths)) {
+                    paths.forEach(path -> {
+                        if (path.toString().endsWith(".json")) {
+                            BlockListManager manager = BlockListHelper.fromJsonPath(world, path);
+                            if (manager == null) return;
+
+                            world.setCurrentlyGeneratingStructureName(() ->
+                                    "\n| ewc multi-chunk feature generating: "
+                                            + "\n\t- step: end generation"
+                                            + "\n\t- center chunkPos: " + chunk.getPos().toString()
+                                            + "\n\t- parent directory: " + path.getParent().getFileName().toString()
+                                            + "\n\t- generating: " + path.getFileName()
+                            );
+                            manager.placeAllNDelete(world);
+                            FileUtil.removeFile(path);
+                        }
+                    });
+                } catch (Exception e) {
+                    e.fillInStackTrace();
+                }
+                try {
+                    Files.delete(basePaths);
+                } catch (IOException e) {
+                    e.fillInStackTrace();
+                }
             }
         }
-        FileUtil.removeGeneratedChunkDirectory(chunk, world);
 
+        Path managerPath = EwcFolderData.getStructureReference(chunk.getPos());
+        try {
+            Files.delete(managerPath);
+        } catch (Exception e) {
+            e.fillInStackTrace();
+        }
     }
 }
